requirements.txt
aiogram==3.4.1
flask[async]==2.3.3
Werkzeug==2.3.7
hypercorn==0.17.3
redis==5.0.1
python-dotenv==1.0.1
rapidfuzz==3.9.6
httpx==0.27.2
pydantic==2.5.3
pydantic-settings==2.2.1
duckduckgo-search==5.3.1
openai==1.51.2
google-generativeai==0.7.2
beautifulsoup4==4.12.3
lxml==5.2.2

google-cloud-vision==3.7.4
pytesseract==0.3.10
Pillow==10.4.0
scikit-learn==1.4.2
joblib==1.4.2
numpy==1.26.4
scipy==1.12.0
